Add files via upload
Program Description:
This project is intended for you to use programming concepts you learned in CSCI 111 (decision
statements, loop statements) and apply them in a simple Java program (using some of the classes
covered in lecture and lab). In addition, you will submit the project through Blackboard to make sure it is
clear how to do that.
We will look at your coding style, documentation (comments) and, of course, that the project works.
Check out the grading scheme in the Projects folder in Blackboard. Ignore Javadoc for this project.
Program Requirements:
Write a Java program that will
1. Ask the user to type in a sentence, using a JOptionPane.showInputDialog().
2. The program will examine each letter in the string and count how many time the upper-case
letter 'E' appears, and how many times the lower-case letter 'e' appears. The key here is to use
the charAt method in class String.
3. Using a JOptionPane.showMessageDialog(), tell the user how many upper and lower case e's
were in the string.
4. Repeat this process until the user types the word "Stop". (Check out the method
equalsIgnoreCase in class String to cover all upper/lower case possibilities of the word
"STOP").